Module bacon_rajan_cc::trace

Tracing traits, types, and implementation.

Traits

Trace
A trait that informs cycle collector how to find memory that is owned by a Trace instance and managed by the cycle collector.

Type Definitions

Tracer
A Tracer is a callback function that is invoked for each CcBoxPtr owned by an instance of something.