Trait bacon_rajan_cc::trace::Trace

pub trait Trace {
    fn trace(&mut self, tracer: &mut Tracer);
}

A trait that informs cycle collector how to find memory that is owned by a Trace instance and managed by the cycle collector.

Required Methods

fn trace(&mut self, tracer: &mut Tracer)

Invoke the Tracer on each of the CcBoxPtrs owned by this Trace instance.

Failing to invoke the tracer on every owned CcBoxPtr can lead to leaking cycles.